otrad: caching logic is wrong with expired headers
This patch cleans up some warnings around logging.

The feed caching isn't expiring correctly using the Expires headers
because HTTP::Date::str2time appears to return the local epoch time
for every string.  Maybe this a raspberry pi thing?

In any case, I need to figure out how to convert the Expires date (RFC
1126) into a unix epoch timestamp.

+++ lib/Otra/App.pm
@@ -1,3 +1,4 @@
+# -*- cperl -*-
 package Otra::App;
 use strict;
 use warnings;
@@ -48,12 +49,14 @@ sub daemonize {
 sub log {
     my ($self) = shift;
 
-    if (-s $self->log_file > 2_000_000) {
-        my $rotate = $self->log_file . ".1";
-        if (-e $rotate) {
-            unlink $rotate;
+    if (-e $self->log_file) {
+        if (-s $self->log_file > 2_000_000) {
+            my $rotate = $self->log_file . ".1";
+            if (-e $rotate) {
+                unlink $rotate;
+            }
+            rename $self->log_file, $rotate;
         }
-        rename $self->log_file, $rotate;
     }
 
     my $msg = sprintf("%s %s\n", scalar(localtime()), join(" ", @_));
